Service layer

+++ src/main/java/com/backend/restro/model/RegistrationRequest.java
@@ -2,10 +2,12 @@ package com.backend.restro.model;
 
 import lombok.Data;
 
+import java.math.BigInteger;
+
 @Data
 public class RegistrationRequest {
     private String name;
     private String email;
     private String password;
-    private Long phoneNumber;
+    private BigInteger phoneNumber;
 }
